+++ src/pharmpy/modeling/plots.py
@@ -1,11 +1,17 @@
+import re
+import warnings
 from typing import List, Optional
 
 import pharmpy.visualization
 from pharmpy.deps import altair as alt
+from pharmpy.deps import numpy as np
 from pharmpy.deps import pandas as pd
-from pharmpy.model import Model
+from pharmpy.deps import scipy, sympy, sympy_stats
+from pharmpy.model import Assignment, Model
 
 from .data import get_observations
+
+norm = scipy.stats.norm
 
 
 def plot_iofv_vs_iofv(iofv1: pd.Series, iofv2: pd.Series, name1: str, name2: str):
@@ -82,3 +88,87 @@ def plot_individual_predictions(
         .facet(f'{idcol}:N', columns=5)
     )
     return plot
+
+
+def plot_transformed_eta_distributions(
+    model: Model, parameter_estimates: dict, individual_estimates: pd.DataFrame
+):
+    """Plot transformed eta distributions for all transformed etas
+
+    Parameters
+    ----------
+    model : Model
+        Previously run Pharmpy model.
+    parameter_estimates : dict
+        Parameter estimates of model fit
+    individual_estimates : pd.DataFrame
+        Individual estimates for etas
+
+    Returns
+    -------
+    alt.Chart
+        Plot
+    """
+    parameter_estimates = dict(parameter_estimates)
+    eta_symbols = {sympy.Symbol(name) for name in model.random_variables.etas.names}
+
+    transformations = []
+    for s in model.statements.before_odes:
+        if isinstance(s, Assignment):
+            m = re.match(r'ETA\w(\d+)', s.symbol.name)
+            if m:
+                symbols = s.expression.free_symbols
+                inter = symbols.intersection(eta_symbols)
+                if len(inter) == 1:
+                    transformations.append((inter.pop(), s.expression))
+
+    x = np.linspace(-2.0, 2.0, 1000)
+    i = 1
+
+    df = pd.DataFrame()
+
+    for eta, expr in transformations:
+        var = model.random_variables.etas.get_covariance(eta.name, eta.name).subs(
+            parameter_estimates
+        )
+        subdf = pd.DataFrame({'x': x, 'original': norm.pdf(x, scale=float(var) ** 0.5)})
+        rv = sympy_stats.Normal('eta', 0, sympy.sqrt(var))
+        expr = expr.subs(parameter_estimates).subs({eta: rv})
+        curdens = sympy_stats.density(expr)
+        densfn = sympy.lambdify(curdens.variables[0], curdens.expr)
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            subdf['boxcox'] = densfn(x)
+        subdf = pd.melt(subdf, id_vars=['x'], value_name='density')
+        eta_name = f'ETA_{i}'
+        ebes = pd.DataFrame(
+            {'x': individual_estimates[eta.name], 'density': 0.0, 'variable': 'ebes'}
+        )
+        subdf = pd.concat([subdf, ebes])
+
+        subdf['eta'] = eta_name
+        df = pd.concat([df, subdf])
+        i += 1
+
+    single = (
+        alt.Chart()
+        .mark_area(opacity=0.3)
+        .encode(
+            x=alt.X('x', axis=alt.Axis(labels=False, ticks=False, title=None)),
+            y='density',
+            color=alt.Color('variable:N', scale=alt.Scale(domain=['original', 'boxcox'])),
+        )
+    )
+    ticks = (
+        alt.Chart()
+        .mark_tick()
+        .encode(
+            x='x',
+            y='density',
+        )
+        .transform_filter(alt.datum.variable == 'ebes')
+    )
+    layer = alt.layer(single, ticks, data=df)
+
+    facet = layer.facet(facet='eta:N', columns=3)
+    return facet